refactor: remove estimates from project store
