+++ web/components/estimates/index.ts
@@ -1,4 +1,5 @@
 export * from "./create-update-estimate-modal";
 export * from "./delete-estimate-modal";
-export * from "./estimate-select";
 export * from "./estimate-list-item";
+export * from "./estimate-select";
+export * from "./estimates-list";
