+++ web/layouts/auth-layout/project-wrapper.tsx
@@ -20,10 +20,11 @@ export const ProjectAuthWrapper: FC<IProjectAuthWrapper> = observer((props) => {
   // store
   const {
     user: { fetchUserProjectInfo, projectMemberInfo, hasPermissionToProject },
-    project: { fetchProjectDetails, fetchProjectEstimates, workspaceProjects },
+    project: { fetchProjectDetails, workspaceProjects },
     projectLabel: { fetchProjectLabels },
     projectMember: { fetchProjectMembers },
     projectState: { fetchProjectStates },
+    projectEstimates: { fetchProjectEstimates },
     cycle: { fetchCycles },
     module: { fetchModules },
     projectViews: { fetchAllViews },
