+++ web/store/project/project-estimates.store.ts
@@ -1,4 +1,4 @@
-import { observable, action, makeObservable, runInAction } from "mobx";
+import { observable, action, makeObservable, runInAction, computed } from "mobx";
 // types
 import { RootStore } from "../root";
 import { IEstimate, IEstimateFormData } from "types";
@@ -9,7 +9,14 @@ export interface IProjectEstimateStore {
   loader: boolean;
   error: any | null;
 
-  // estimates
+  // observables
+  estimates: {
+    [projectId: string]: IEstimate[] | null; // project_id: members
+  } | null;
+
+  // actions
+  getProjectEstimateById: (estimateId: string) => IEstimate | null;
+  fetchProjectEstimates: (workspaceSlug: string, projectId: string) => Promise<void>;
   createEstimate: (workspaceSlug: string, projectId: string, data: IEstimateFormData) => Promise<IEstimate>;
   updateEstimate: (
     workspaceSlug: string,
@@ -18,14 +25,23 @@ export interface IProjectEstimateStore {
     data: IEstimateFormData
   ) => Promise<IEstimate>;
   deleteEstimate: (workspaceSlug: string, projectId: string, estimateId: string) => Promise<void>;
+
+  // computed
+  projectEstimates: IEstimate[] | undefined;
 }
 
 export class ProjectEstimatesStore implements IProjectEstimateStore {
   loader: boolean = false;
   error: any | null = null;
 
+  // observables
+  estimates: {
+    [projectId: string]: IEstimate[]; // projectId: estimates
+  } | null = {};
+
   // root store
   rootStore;
+
   // service
   projectService;
   estimateService;
@@ -36,16 +52,60 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
       loader: observable,
       error: observable,
 
-      // estimates
+      estimates: observable.ref,
+
+      // actions
+      getProjectEstimateById: action,
+      fetchProjectEstimates: action,
       createEstimate: action,
       updateEstimate: action,
       deleteEstimate: action,
+
+      // computed
+      projectEstimates: computed,
     });
 
     this.rootStore = _rootStore;
     this.projectService = new ProjectService();
     this.estimateService = new ProjectEstimateService();
   }
+
+  get projectEstimates() {
+    const projectId = this.rootStore.project.projectId;
+
+    if (!projectId) return undefined;
+    return this.estimates?.[projectId] || undefined;
+  }
+
+  getProjectEstimateById = (estimateId: string) => {
+    const estimates = this.projectEstimates;
+    if (!estimates) return null;
+    const estimateInfo: IEstimate | null = estimates.find((estimate) => estimate.id === estimateId) || null;
+    return estimateInfo;
+  };
+
+  fetchProjectEstimates = async (workspaceSlug: string, projectId: string) => {
+    try {
+      this.loader = true;
+      this.error = null;
+
+      const estimatesResponse = await this.estimateService.getEstimatesList(workspaceSlug, projectId);
+      const _estimates = {
+        ...this.estimates,
+        [projectId]: estimatesResponse,
+      };
+
+      runInAction(() => {
+        this.estimates = _estimates;
+        this.loader = false;
+        this.error = null;
+      });
+    } catch (error) {
+      console.error(error);
+      this.loader = false;
+      this.error = error;
+    }
+  };
 
   createEstimate = async (workspaceSlug: string, projectId: string, data: IEstimateFormData) => {
     try {
@@ -62,9 +122,9 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
       };
 
       runInAction(() => {
-        this.rootStore.project.estimates = {
-          ...this.rootStore.project.estimates,
-          [projectId]: [responseEstimate, ...(this.rootStore.project.estimates?.[projectId] || [])],
+        this.estimates = {
+          ...this.estimates,
+          [projectId]: [responseEstimate, ...(this.estimates?.[projectId] || [])],
         };
       });
 
@@ -76,12 +136,12 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
   };
 
   updateEstimate = async (workspaceSlug: string, projectId: string, estimateId: string, data: IEstimateFormData) => {
-    const originalEstimates = this.rootStore.project.getProjectEstimateById(estimateId);
+    const originalEstimates = this.getProjectEstimateById(estimateId);
 
     runInAction(() => {
-      this.rootStore.project.estimates = {
-        ...this.rootStore.project.estimates,
-        [projectId]: (this.rootStore.project.estimates?.[projectId] || [])?.map((estimate) =>
+      this.estimates = {
+        ...this.estimates,
+        [projectId]: (this.estimates?.[projectId] || [])?.map((estimate) =>
           estimate.id === estimateId ? { ...estimate, ...data.estimate } : estimate
         ),
       };
@@ -95,15 +155,15 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
         data,
         this.rootStore.user.currentUser!
       );
-      await this.rootStore.project.fetchProjectEstimates(workspaceSlug, projectId);
+      await this.fetchProjectEstimates(workspaceSlug, projectId);
 
       return response;
     } catch (error) {
       console.log("Failed to update estimate from project store");
       runInAction(() => {
-        this.rootStore.project.estimates = {
-          ...this.rootStore.project.estimates,
-          [projectId]: (this.rootStore.project.estimates?.[projectId] || [])?.map((estimate) =>
+        this.estimates = {
+          ...this.estimates,
+          [projectId]: (this.estimates?.[projectId] || [])?.map((estimate) =>
             estimate.id === estimateId ? { ...estimate, ...originalEstimates } : estimate
           ),
         };
@@ -113,14 +173,12 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
   };
 
   deleteEstimate = async (workspaceSlug: string, projectId: string, estimateId: string) => {
-    const originalEstimateList = this.rootStore.project.projectEstimates || [];
+    const originalEstimateList = this.projectEstimates || [];
 
     runInAction(() => {
-      this.rootStore.project.estimates = {
-        ...this.rootStore.project.estimates,
-        [projectId]: (this.rootStore.project.estimates?.[projectId] || [])?.filter(
-          (estimate) => estimate.id !== estimateId
-        ),
+      this.estimates = {
+        ...this.estimates,
+        [projectId]: (this.estimates?.[projectId] || [])?.filter((estimate) => estimate.id !== estimateId),
       };
     });
 
@@ -131,8 +189,8 @@ export class ProjectEstimatesStore implements IProjectEstimateStore {
       console.log("Failed to delete estimate from project store");
       // reverting back to original estimate list
       runInAction(() => {
-        this.rootStore.project.estimates = {
-          ...this.rootStore.project.estimates,
+        this.estimates = {
+          ...this.estimates,
           [projectId]: originalEstimateList,
         };
       });
