+++ web/store/project/project.store.ts
@@ -1,9 +1,9 @@
 import { observable, action, computed, makeObservable, runInAction } from "mobx";
 // types
 import { RootStore } from "../root";
-import { IProject, IEstimate } from "types";
+import { IProject } from "types";
 // services
-import { ProjectService, ProjectStateService, ProjectEstimateService } from "services/project";
+import { ProjectService, ProjectStateService } from "services/project";
 import { IssueService, IssueLabelService } from "services/issue";
 
 export interface IProjectStore {
@@ -16,14 +16,10 @@ export interface IProjectStore {
   project_details: {
     [projectId: string]: IProject; // projectId: project Info
   };
-  estimates: {
-    [projectId: string]: IEstimate[] | null; // project_id: members
-  } | null;
 
   // computed
   searchedProjects: IProject[];
   workspaceProjects: IProject[] | null;
-  projectEstimates: IEstimate[] | null;
   joinedProjects: IProject[];
   favoriteProjects: IProject[];
   currentProjectDetails: IProject | undefined;
@@ -34,10 +30,8 @@ export interface IProjectStore {
 
   getProjectById: (workspaceSlug: string, projectId: string) => IProject | null;
 
-  getProjectEstimateById: (estimateId: string) => IEstimate | null;
   fetchProjects: (workspaceSlug: string) => Promise<void>;
   fetchProjectDetails: (workspaceSlug: string, projectId: string) => Promise<any>;
-  fetchProjectEstimates: (workspaceSlug: string, projectId: string) => Promise<any>;
 
   addProjectToFavorites: (workspaceSlug: string, projectId: string) => Promise<any>;
   removeProjectFromFavorites: (workspaceSlug: string, projectId: string) => Promise<any>;
@@ -62,9 +56,6 @@ export class ProjectStore implements IProjectStore {
   project_details: {
     [projectId: string]: IProject; // projectId: project
   } = {};
-  estimates: {
-    [projectId: string]: IEstimate[]; // projectId: estimates
-  } | null = {};
 
   // root store
   rootStore;
@@ -73,7 +64,6 @@ export class ProjectStore implements IProjectStore {
   issueLabelService;
   issueService;
   stateService;
-  estimateService;
 
   constructor(_rootStore: RootStore) {
     makeObservable(this, {
@@ -86,13 +76,9 @@ export class ProjectStore implements IProjectStore {
       projects: observable.ref,
       project_details: observable.ref,
 
-      estimates: observable.ref,
-
       // computed
       searchedProjects: computed,
       workspaceProjects: computed,
-
-      projectEstimates: computed,
 
       currentProjectDetails: computed,
 
@@ -106,9 +92,6 @@ export class ProjectStore implements IProjectStore {
       fetchProjectDetails: action,
 
       getProjectById: action,
-      getProjectEstimateById: action,
-
-      fetchProjectEstimates: action,
 
       addProjectToFavorites: action,
       removeProjectFromFavorites: action,
@@ -125,7 +108,6 @@ export class ProjectStore implements IProjectStore {
     this.issueService = new IssueService();
     this.issueLabelService = new IssueLabelService();
     this.stateService = new ProjectStateService();
-    this.estimateService = new ProjectEstimateService();
   }
 
   get searchedProjects() {
@@ -162,11 +144,6 @@ export class ProjectStore implements IProjectStore {
   get favoriteProjects() {
     if (!this.rootStore.workspace.workspaceSlug) return [];
     return this.projects?.[this.rootStore.workspace.workspaceSlug]?.filter((p) => p.is_favorite);
-  }
-
-  get projectEstimates() {
-    if (!this.projectId) return null;
-    return this.estimates?.[this.projectId] || null;
   }
 
   // actions
@@ -221,37 +198,6 @@ export class ProjectStore implements IProjectStore {
 
     const projectInfo: IProject | null = projects.find((project) => project.id === projectId) || null;
     return projectInfo;
-  };
-
-  getProjectEstimateById = (estimateId: string) => {
-    if (!this.projectId) return null;
-    const estimates = this.projectEstimates;
-    if (!estimates) return null;
-    const estimateInfo: IEstimate | null = estimates.find((estimate) => estimate.id === estimateId) || null;
-    return estimateInfo;
-  };
-
-  fetchProjectEstimates = async (workspaceSlug: string, projectId: string) => {
-    try {
-      this.loader = true;
-      this.error = null;
-
-      const estimatesResponse = await this.estimateService.getEstimatesList(workspaceSlug, projectId);
-      const _estimates = {
-        ...this.estimates,
-        [projectId]: estimatesResponse,
-      };
-
-      runInAction(() => {
-        this.estimates = _estimates;
-        this.loader = false;
-        this.error = null;
-      });
-    } catch (error) {
-      console.error(error);
-      this.loader = false;
-      this.error = error;
-    }
   };
 
   addProjectToFavorites = async (workspaceSlug: string, projectId: string) => {
